feat: refactor loader
