+++ frontend/src/views/PoemCreateView.vue
@@ -2,7 +2,7 @@
 
 import { ref } from 'vue'
 import { useAppStore } from '@/stores/app'
-import Loader from '../components/Loader.vue'
+import Loader from '../components/CustomLoader.vue'
 import AcNotification from '../components/ac-notification.vue'
 import { useRouter } from 'vue-router'
 
@@ -105,7 +105,7 @@ const submitPoem = async () => {
               class='flex w-full justify-center rounded-md bg-emerald-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-emerald-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-emerald-600'
               data-test='save-poem-button'
               type='submit'>
-            <Loader v-if='loading' class='mr-6' />
+            <CustomLoader v-if='loading' class='mr-6' />
             Create Poem
           </button>
         </div>
